requirements.txt
discord.py
aiohttp
lxml
beautifulsoup4
requests